poetry_ml/bank/pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = "Direct Marketing Project"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"
scikit-learn = "^1.0"
typer = {extras = ["all"], version = "^0.4.0"}
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"
pre-commit = "^2.15.0"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
notebook = "^6.4.4"
ipykernel = "^6.4.1"
black = "^21.9b0"
pep8 = "^1.7.1"
flake8 = "^4.0.1"
Sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
mypy = "^0.910"
pep8-naming = "^0.12.1"
pytest = "^6.2.5"

[tool.poetry.scripts]
bank = "bank.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
